+++ server/core/gateway.cc
@@ -427,6 +427,15 @@ static void sigfatal_handler(int i)
               cnf->sysname,
               cnf->release_string);
 
+    if (DCB* dcb = dcb_get_current())
+    {
+        if (dcb->session)
+        {
+            session_dump_statements(dcb->session);
+            session_dump_log(dcb->session);
+        }
+    }
+
     auto cb = [](const char* symbol, const char* cmd) {
             MXS_ALERT("  %s: %s", symbol, cmd);
         };